re-factor of Hero load animation
