+++ src/index.html
@@ -14,7 +14,7 @@
   </head>
   <body>
     <header>
-      <div class="header-overlay">
+      <!-- <div class="header-overlay">
         <div class="title">
           <p>USER_</p>
           <div class="title-user">
@@ -45,23 +45,44 @@
               <textPath xlink:href="#text-path">Scroll Down</textPath>
             </text>
           </svg>
-          <div class="header-links-container">
-            <button id="about-link" class="header-links">
-              <span>ABOUT_</span>
-            </button>
-            <button id="projects-link" class="header-links">
-              <span>PROJECTS_</span>
-            </button>
-            <button id="contact-link" class="header-links">
-              <span>CONTACT_</span>
-            </button>
-            <button id="contact-link" class="header-links">
-              <a href="https://drive.google.com/file/d/1Px8X1X0ivDgjOzrM0GXUYN-iA_sGuBmP/view?usp=sharing" target="_blank">RESUME_</a>
-            </button>
-          </div>
         </div>
       </div>
+    </div> -->
+
+    <div class="header-links-container">
+      <button id="about-link" class="header-links">
+        <span>About+</span>
+      </button>
+      <button id="projects-link" class="header-links">
+        <span>PROJECTS+</span>
+      </button>
+      <button id="contact-link" class="header-links">
+        <span>CONTACT+</span>
+      </button>
+      <button id="contact-link" class="header-links">
+        <a href="https://drive.google.com/file/d/1Px8X1X0ivDgjOzrM0GXUYN-iA_sGuBmP/view?usp=sharing" target="_blank">RESUME+</a>
+      </button>
+    </div>
+    <canvas></canvas>
+
+      <div class="login-container">
+        <h2>LOGIN_</h2>
+        <div>
+          <p>User: <span id="login-user"></span><span class="hero-cursor" /></p>
+          <p id="login-pass-field">Password: <span id="login-pass"></span></p>
+        </div>
+        <div id="login-execute">EXECUTE_
+          <section id="login-execute-overlay"></section>
+        </div>
+      </div>
+
+      <div class="header-user">
+        <p>User_</p>
+        <h1>APPLEFRITTR</h1>
+      </div>
     </header>
+
+
     <main class="about">
       <div class="section-heading">
         <h2 class="header" id="about-head">ABOUT_</h2>
@@ -462,9 +483,9 @@
         </ul>
         <ul class="footer-links-list">
           <p>Project Links</p>
+          <li><a href="https://github.com/Applefrittr/bust-a-move" target="_blank">Bust A Move</a></li>
           <li><a href="https://discognito-app-production.up.railway.app" target="_blank">Discognito</a></li>
           <li><a href="https://github.com/Applefrittr/tarkov-loadouts" target="_blank">Tarkov Loadouts</a></li>
-          <li><a href="https://github.com/Applefrittr/wheres-waldo" target="_blank">Pokefind</a></li>
           <li><a href="https://github.com/Applefrittr/messenger-app" target="_blank">Mylo Messenger</a></li>
           <li><a href="https://applefrittr.github.io/space-invaders/" target="_blank">Space Invaders</a></li>
         </ul>
