+++ src/styles/hero.css
@@ -1,0 +1,195 @@
+canvas {
+  position: absolute;
+  left: 0;
+  top: 0;
+  z-index: 10;
+}
+
+.login-container {
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+  width: 300px;
+
+  h2 {
+    background-color: var(--bg-color);
+    font-family: Tomorrow;
+  }
+
+  div {
+    display: flex;
+    flex-direction: column;
+    gap: 0.5rem;
+
+    p {
+      height: 20px;
+      display: flex;
+      align-items: center;
+    }
+  }
+
+  h2,
+  div {
+    padding: 0.5rem 1rem;
+    border: 1px solid var(--lettering);
+  }
+
+  #login-execute {
+    width: max-content;
+    position: relative;
+    font-family: Tomorrow;
+    font-weight: bolder;
+    transition: background-color 0.25s ease-in-out;
+
+    #login-execute-overlay {
+      position: absolute;
+      left: -1px;
+      top: -1px;
+      width: 105%;
+      height: 110%;
+      background-color: black;
+      transition: transform 1s ease-in-out;
+    }
+  }
+}
+
+#login-user,
+#login-pass {
+  padding-left: 0.5rem;
+}
+
+.hero-cursor {
+  height: inherit;
+  width: 14px;
+  display: inline-block;
+  animation: blink 1s linear infinite;
+}
+
+.header-user {
+  position: absolute;
+  top: 20%;
+  left: 20%;
+  display: none;
+  z-index: 15;
+  flex-direction: column;
+  gap: 1rem;
+  opacity: 0;
+  transition: opacity 1s ease-in-out;
+
+  h1,
+  p {
+    background-color: rgba(0, 0, 0, 0.75);
+    border: 1px solid var(--lettering);
+    padding: 1rem 2rem;
+  }
+
+  h1 {
+    font-family: Glitch;
+    font-size: 5rem;
+  }
+
+  p {
+    width: max-content;
+    background-color: var(--bg-color);
+  }
+}
+
+.header-links-container {
+  position: absolute;
+  right: 10%;
+  bottom: 10%;
+  display: none;
+  gap: 1rem;
+  flex-direction: column;
+  opacity: 0;
+  transition: opacity 1s ease;
+  z-index: 15;
+
+  .header-links {
+    background-color: transparent;
+    border: none;
+    outline: none;
+    font-size: 2rem;
+    text-align: left;
+    font-family: Tomorrow;
+    font-weight: bolder;
+    clip-path: polygon(92% 0, 100% 25%, 100% 100%, 8% 100%, 0% 75%, 0 0);
+    background-color: var(--lettering);
+    padding: 1rem;
+    position: relative;
+    width: 230px;
+    height: 60px;
+    border: 0;
+
+    a,
+    span {
+      display: flex;
+      align-items: center;
+      justify-content: center;
+      position: absolute;
+      color: var(--lettering);
+      top: 2px;
+      left: 2px;
+      right: 2px;
+      bottom: 2px;
+      background-color: rgba(0, 0, 0, 0.75);
+      clip-path: polygon(92% 0, 100% 25%, 100% 100%, 8% 100%, 0% 75%, 0 0);
+      text-decoration: none;
+    }
+  }
+}
+
+.login-container::before,
+.header-user::before,
+.header-links-container::before {
+  content: "+";
+}
+
+.login-container::after,
+.header-user::after,
+.header-links-container::after {
+  content: "+";
+  text-align: right;
+}
+
+/* .header-links > span,
+.header-links > a {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  position: absolute;
+  color: var(--lettering);
+  top: 2px;
+  left: 2px;
+  right: 2px;
+  bottom: 2px;
+  background-color: rgba(0, 0, 0);
+  clip-path: polygon(92% 0, 100% 25%, 100% 100%, 8% 100%, 0% 75%, 0 0);
+  text-decoration: none;
+} */
+
+.header-links:hover {
+  cursor: pointer;
+}
+
+.header-links > span:hover,
+.header-links > a:hover {
+  background-color: rgba(0, 0, 0);
+}
+
+.scroll-svg {
+  opacity: 0;
+  transition: opacity 1s ease;
+}
+
+.slide {
+  transform: translateX(100%);
+}
+
+.reveal {
+  display: flex;
+}
+
+.fade-in {
+  opacity: 1;
+}
