+++ src/styles/layout.css
@@ -12,6 +12,9 @@ body {
 header {
   height: 100vh;
   display: flex;
+  justify-content: center;
+  align-items: center;
+  position: relative;
 }
 
 .header-overlay {
@@ -28,64 +31,6 @@ header {
   position: absolute;
   right: 30px;
   bottom: 30px;
-}
-
-.header-links-container {
-  position: absolute;
-  right: 20px;
-  bottom: 150px;
-  display: flex;
-  gap: 1rem;
-  flex-direction: column;
-  opacity: 0;
-  transition: opacity 1s ease;
-}
-
-.header-links {
-  background-color: transparent;
-  border: none;
-  outline: none;
-  font-size: 2rem;
-  text-align: left;
-  font-family: Tomorrow;
-  font-weight: bold;
-  clip-path: polygon(92% 0, 100% 25%, 100% 100%, 8% 100%, 0% 75%, 0 0);
-  background-color: var(--outlines);
-  padding: 1rem;
-  position: relative;
-  width: 230px;
-  height: 60px;
-  border: 0;
-}
-
-.header-links > span,
-.header-links > a {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  position: absolute;
-  color: white;
-  top: 2px;
-  left: 2px;
-  right: 2px;
-  bottom: 2px;
-  background-color: var(--theme-color);
-  clip-path: polygon(92% 0, 100% 25%, 100% 100%, 8% 100%, 0% 75%, 0 0);
-  text-decoration: none;
-}
-
-.header-links:hover {
-  cursor: pointer;
-}
-
-.header-links > span:hover,
-.header-links > a:hover {
-  background-color: var(--outlines);
-}
-
-.scroll-svg {
-  opacity: 0;
-  transition: opacity 1s ease;
 }
 
 .title {
