+++ webpack.config.js
@@ -1,10 +1,9 @@
 const path = require("path");
-const HtmlWebpackPlugin = require('html-webpack-plugin');
-
+const HtmlWebpackPlugin = require("html-webpack-plugin");
 
 module.exports = {
   mode: "development",
-  entry: "./src/index.js",
+  entry: "./src/js/index.js",
   devtool: "inline-source-map",
   devServer: {
     static: "./dist",
@@ -12,15 +11,15 @@ module.exports = {
   plugins: [
     new HtmlWebpackPlugin({
       title: "Personal Portfolio",
-      inject: 'body',
-      template: './src/index.html',
-      filename: 'index.html'
+      inject: "body",
+      template: "./src/index.html",
+      filename: "index.html",
     }),
   ],
   output: {
     filename: "main.js",
     path: path.resolve(__dirname, "dist"),
-    assetModuleFilename: 'assets/[hash][ext][query]'
+    assetModuleFilename: "assets/[hash][ext][query]",
   },
   module: {
     rules: [
